Rope works, calculated points are preserved.
