+++ src/mesh/index.js
@@ -7,4 +7,4 @@ export { default as CanvasMeshRenderer } from './canvas/CanvasMeshRenderer';
 export { default as Plane } from './Plane';
 export { default as NineSlicePlane } from './NineSlicePlane';
 export { default as Rope } from './Rope';
-export { default as RopeSprite } from './RopeSprite';
+export { default as RopePoint } from './RopePoint';
